took everything out of main and now its ready for new features
