+++ Pokemon/Game.h
@@ -1,0 +1,9 @@
+#pragma once
+
+class Player;
+class Game
+{
+public:
+	Game();
+	void gameLoop(Player& player);
+};
